Organize project structure into packages
