+++ src/main/java/com/example/users/model/User.java
@@ -1,4 +1,4 @@
-package com.example.users;
+package com.example.users.model;
 
 import jakarta.persistence.*;
 import lombok.Data;
