+++ src/main/java/com/example/users/producer/UserCacheEvictProducer.java
@@ -1,4 +1,4 @@
-package com.example.users;
+package com.example.users.producer;
 
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
